pyproject.toml
[project]
name = "triagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "crewai>=0.51.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
]
